refactor: restructure code directory
